test(manager-api): wide range of notification and event tests

+++ manager/api/beans/src/main/java/io/apiman/manager/api/beans/events/ApimanEvent.java
@@ -23,5 +23,5 @@ public @interface ApimanEvent {
     /**
      * Version of the event
      */
-    int version();
+    long version();
 }
